feat: add first section of landing page
Completed the layout for the navigation bar and the hero section
using flexbox.
The items on the left have the same width (400px) and
the items on right have the same width (500px). This will make them
align on the vertical lign.
The unordered list flex items are positioned at the flex-end to match
the right line of the image container.